Moving ingestor integration test to be a FeatureTest instead
